dp/product & gp/product
This commit includes two updates. The first is to allow for dp/product
and gp/product Amazon URLs. I happened to just bring up a recent item in
my history and it had a gp/product in it instead of a dp/product:

    https://www.amazon.com/gp/product/0307336794

vs

    https://www.amazon.com/dp/product/0307336794

They both seem to work so I adjusted scrape.py and carousel.js to allow
for either.

The first time I ran the program I got a file not found error because
scrape.py was trying to write to a non-existent results/ directory.
I like the --prefix option for recording the provenance of where
the data came from (linked to the seed file. So I made some adjustments
to make that a bit clearer, and also to create the directory if
instructed to write somewhere that doesn't exist yet.